start_slackbot/requirements.txt
slackbot
python-dotenv
